requirements.txt
python-binance==1.0.16
pandas==1.4.4
numpy==1.23.3
matplotlib==3.6.0
ta==0.10.2
python-dotenv==0.21.0
schedule==1.1.0
requests==2.28.1
talib-binary==0.4.19
